Keep compiler happy

+++ Source/FLIMReader/AbstractFifoReader.h
@@ -131,7 +131,7 @@ public:
 
    void alignFrames();
 
-   float getProgress() { return event_reader->getProgress(); }
+   double getProgress() { return event_reader->getProgress(); }
 
    void setTemporalResolution(int temporal_resolution);
